requirements.txt
anyio==3.6.2
attrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
click==8.1.3
colorama==0.4.6
cryptography==38.0.4
exceptiongroup==1.1.0
fastapi==0.88.0
greenlet==2.0.1
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.1
idna==3.4
iniconfig==1.1.1
jedi==0.18.2
jwt==1.3.1
msgpack==1.0.4
packaging==22.0
parso==0.8.3
passlib==1.7.4
pluggy==1.0.0
pycparser==2.21
pydantic==1.10.2
pynvim==0.4.3
pytest==7.2.0
python-dotenv==0.21.0
PyYAML==6.0
rfc3986==1.5.0
sniffio==1.3.0
SQLAlchemy==1.4.44
starlette==0.22.0
tomli==2.0.1
typing_extensions==4.4.0
uvicorn==0.20.0
watchfiles==0.18.1
websockets==10.4
